fix: alias imports - rename all the component files to `index.tsx` in their respecitve directories - add direct alias imports where it was being us indirectly
